chore: separate pages for role creation and editing
